backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
pdfplumber
httpx
python-multipart
supabase
python-jose[cryptography]
python-dotenv
fastapi
uvicorn
gunicorn
python-multipart
flask-cors
